backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
python-decouple
gunicorn
pytest
pytest-django
djoser
yfinance
celery==5.3.4
django-celery-beat
